FIX: Add new authentication middleware
Remove bootstrap system, it was causing problems when the database
didn't exist, for example when running

    $ bundle exec rake db:create

To use the default system admin I add a new authentication middleware,
Guard, that sets the admin user on the fly if needed. It also sets
the authenticated user in `env["guard.user"]` so the Sinatra app can
use it directly.