FIX: Ensure thread safety on Guard middleware
If the app receives lots of request using a thread based server like
puma, there is a possibility that the users of each request get crossed.

For example, a guest user trying to do a POST call could do it because
during its request time the system authenticate it like if it was an
admin user doing another request at the same time.

This commit fix this by calling `#dup` inside the `Guard#call` method.
This ensures that each request will be handled by a new instance with a
copy of the Rack environment, preventing sharing global variables
accross threads. This is the same thing that Sinatra does.